Description: Fixed UI delay issues when getting drinks menu of new venue for the first time.
Issue: To reproduce the issue without the fix, reset all application data and check in a new venue. Displaying the list of drinks takes an excessive amount of time, sometimes over 10 seconds.

Fix: To solve this issue we introduce menu caching. We create a new data structure in the Drinks fragment to store the menu. We also provide a smart loader that decides whether or not to display the drinks from cache, fetch them from the DB or if they don't exist in the DB get them from the host. The loader fetches the drinks from the host in memory and 

Testing: With this fix, displaying the menu for a venue to which we check in for the first time should take no more than a second instead of over 10 seconds.

Known issues: Notice that the menu is saved in the DB permanently, which is not necessarily correct because the menu may change periodically. I'm not yet sure why a database in used here, so let's think about this and see if it makes sense to keep it.

git-svn-id: http://192.168.2.90/svn/Vendsy/Bartsy/Android/Bartsy@328 c3b44e55-a100-42f6-9c86-32dbf579054b